requirements.txt
flask
flask-cors
tensorflow
transformers
opencv-python-headless
numpy
Pillow
matplotlib
